Add files via upload
HTML/CSS/JS one page ACME Wallet website with script, schematic, logo banner (white bg), and download button under logo. The SCHEMATIC IS A TEMP PLACEHOLDER and will change with the correct information displayed. I created a backup to the index and CSS files in case of emergency just swap them out, the CSS is not SCSS and all edits can be destructive. Colors, fonts, and style reflect the logo Albatross and I came up with. This is a one page website with a nav bar that sends you down to the sections instead of a different page. The Chrome plugin download button is a placeholder to see how it would look and if we are going to offer a download for it.